pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "data-contract-validator"
version = "0.1.0"
description = "A small Python project to validate CSV and JSON data contracts"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0.0",
    "email-validator>=2.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.6.0",
    "python-dotenv>=1.0.1",
]

[project.scripts]
data-contract-validator = "data_contract_validator.cli:main"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

